+++ magick/quantum-import.c
@@ -212,7 +212,7 @@ static inline const unsigned char *PushQuantumPixel(
     quantum_bits=(size_t) i;
     if (quantum_bits > quantum_state->bits)
       quantum_bits=quantum_state->bits;
-    i-=quantum_bits;
+    i-=(ssize_t) quantum_bits;
     quantum_state->bits-=quantum_bits;
     *quantum=(*quantum << quantum_bits) | ((quantum_state->pixel >>
       quantum_state->bits) &~ ((~0UL) << quantum_bits));
@@ -244,7 +244,7 @@ static inline const unsigned char *PushQuantumLongPixel(
       quantum_bits=quantum_state->bits;
     *quantum|=(((quantum_state->pixel >> (32UL-quantum_state->bits)) &
       quantum_state->mask[quantum_bits]) << (depth-i));
-    i-=quantum_bits;
+    i-=(ssize_t) quantum_bits;
     quantum_state->bits-=quantum_bits;
   }
   return(pixels);
